Last refactor
