+++ db_handler/db_handler/model/query_request.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, Field
 from typing import Optional
 
-from db_handler.model.sql_condition_group import SqlConditionGroup
+from db_handler.db_handler.model.sql_condition_group import SqlConditionGroup
 
 
 class QueryRequest(BaseModel):
