+++ db_handler/db_handler/model/sql_condition.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 from typing import Any
 
-from db_handler.model.type.sql_condition_operator import SqlConditionOperator
+from db_handler.db_handler.model.type.sql_condition_operator import SqlConditionOperator
 
 
 class SqlCondition(BaseModel):
