--- db_handler/model/sql_condition_group.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-from db_handler.model.sql_condition import SqlCondition
-from db_handler.model.type.sql_join import SqlJoin
-
-
-class SqlConditionGroup(BaseModel):
-    conditions: list[SqlCondition]
-    join: SqlJoin
