+++ requirements.txt
@@ -1,2 +1,3 @@
 psycopg2-binary==2.9.10
 pydantic==2.10.6
+pytest==8.3.4
